crimedata: a package for accessing US city crime data

Access incident-level crime data from the Open Crime Database

Crime Open Database

The Crime Open Database (CODE) is a service that makes it convenient to use crime data from multiple US cities in research on crime. All the data are available to use for free as long as you acknowledge the source of the data.

For more about CODE data, see https://osf.io/zyaqn/.

Accessing the data

To access CODE data, call get_crime_data. Data are returned as a 'tidy' tibble with each row corresponding to one recorded crime.

Chicago data license

This site provides applications using data that has been modified for use from its original source, https://www.chicago.gov/, the official website of the City of Chicago. The City of Chicago makes no claims as to the content, accuracy, timeliness, or completeness of any of the data provided at this site. The data provided at this site is subject to change at any time. It is understood that the data provided at this site is being used at one's own risk.

See also

Author

Maintainer: Matthew Ashby matthew.ashby@ucl.ac.uk (ORCID) [copyright holder]