Thefts of motor vehicles 2014 to 2017

Dataset containing records of thefts of motor vehicles in New York City from 2014 to 2017, obtained from the Crime Open Database.

Usage

nycvehiclethefts

Format

A tibble with 35,746 rows and 13 variables:

uid

an integer unique identifier for the offense

date_single

date (and, in most cases, time) half-way between the first and last possible dates at which the offense could have occurred

date_start

first possible date (and, in most cases, time) at which the offense could have occurred

date_send

last possible date (and, in most cases, time) at which the offense could have occurred

longitude

approximate longitude

latitude

approximate latitude

location_type

type of location*

location_category

category of location type*

fips_state

two-digit FIPS state code (possibly with leading zero)

fips_county

three-digit FIPS county code (possibly with leading zero)

tract

six-digit code for 2016 census tract

block_group

one-digit code for 2016 census block group

block

four-digit code for 2016 census block

Source

https://osf.io/zyaqn/

Details

More details of the data format are available on the Crime Open Database website. Variables marked * are only available for some of the data, due to limitations in the data published by some cities.

The variables in this dataset mirror those obtained by calling get_crime_data(type = "core"), except that some fields have been removed because they are redundant (e.g. if they have the same value for all rows in this dataset).